refactor: facade에서 domainservice 구조 변경 (#53)
