+++ src/main/java/com/aztgg/api/auth/presentation/AuthController.java
@@ -1,11 +1,9 @@
 package com.aztgg.api.auth.presentation;
 
-import com.aztgg.api.auth.application.AuthService;
+import com.aztgg.api.auth.application.UserFacadeService;
 import com.aztgg.api.auth.application.dto.request.AdminLoginRequest;
 import com.aztgg.api.auth.application.dto.request.KakaoLoginRequest;
 import com.aztgg.api.auth.application.dto.response.LoginResponse;
-import com.aztgg.api.auth.application.strategy.impl.AdminAuthCredentials;
-import com.aztgg.api.auth.application.strategy.impl.KakaoAuthCredentials;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -15,32 +13,30 @@ import org.springframework.web.bind.annotation.*;
 @RequestMapping("/v1/auth")
 public class AuthController implements AuthApi {
 
-    private final AuthService authService;
+    private final UserFacadeService userFacadeService;
 
     @Override
     @PostMapping("/admin/login")
     public ResponseEntity<LoginResponse> adminLogin(@RequestBody AdminLoginRequest loginRequest) {
-        AdminAuthCredentials credentials = AdminAuthCredentials.of(loginRequest.username(), loginRequest.password());
-        return ResponseEntity.ok(authService.authenticate(credentials));
+        return ResponseEntity.ok(userFacadeService.adminLogin(loginRequest));
     }
 
     @Override
     @PostMapping("/kakao/login")
     public ResponseEntity<LoginResponse> kakaoLogin(@RequestBody KakaoLoginRequest loginRequest) {
-        KakaoAuthCredentials credentials = KakaoAuthCredentials.of(loginRequest.kakaoToken());
-        return ResponseEntity.ok(authService.authenticate(credentials));
+        return ResponseEntity.ok(userFacadeService.kakaoLogin(loginRequest));
     }
 
     @Override
     @PostMapping("/refresh")
     public ResponseEntity<LoginResponse> refresh(@CookieValue("refreshToken") String refreshToken) {
-        return ResponseEntity.ok(authService.refresh(refreshToken));
+        return ResponseEntity.ok(userFacadeService.refresh(refreshToken));
     }
 
     @Override
     @PostMapping("/logout")
     public ResponseEntity<Void> logout(@CookieValue("refreshToken") String refreshToken) {
-        authService.logout(refreshToken);
+        userFacadeService.logout(refreshToken);
         return ResponseEntity.noContent().build();
     }
 }
